requirements.txt
Django==1.6.5
Unipath==1.0
djangorestframework==2.3.14
